[hash.c] Resizes now preserve the _initial_size of the hash table type.
This means that you can't resize smaller than the original size of the
hash table type when it was initialized.

Also removed the SIM_HASH_INITIAL_SIZE check, as it is no longer
relevant when it comes to resizing. Instead, the size on hash table type
initialization checks the size passed in to it and sees if it's smaller
than the renamed "SIM_HASH_DEFAULT_SIZE".